+++ src/lib.rs
@@ -181,7 +181,7 @@ crate, add it as a dependency inside [*Cargo.toml*]:
 fixed = "1.11"
 ```
 
-The *fixed* crate requires rustc version 1.53.0 or later.
+The *fixed* crate requires rustc version 1.57.0 or later.
 
 ## Optional features